// src/app/master/models/chemcatgm.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class ChemCatgm {

    chem_pkid: string;
    chem_ritc_id: string;
    chem_ritc_code: string;
    chem_ritc_name:string;
    chem_code: string;
    chem_desc: string;

    rec_locked: boolean;
    rec_mode: string;

    _globalvariables: GlobalVariables;

}